omczmq: change minimum libczmq requirement again
On Ubuntu 16, we do not have libczmq 4, but omczmq obviously builds
fine with 3.0.2, so this seems to be the actual requirement. If that
works, I'll settle for now as I am finally totally confused on what
happens here. I have asked for help from those in the know, but
if we do not get a definitive answer, it'll stay so that it works
for us -- after all this is a contributed module...

closes https://github.com/rsyslog/rsyslog/issues/872