begin documentation on rsyslog containers
The idea is to provide useful information as early as possible and
extend it as it grows. While this at first may look sparse, I think
it is the best way forward. The alternative would be to wait a
long time until we know "everything" and can document it.

see also https://github.com/rsyslog/rsyslog-doc/issues/491