Update build process to default to static value builds
Instead of assuming that we are always building from
the master branch or a feature branch, placeholders have been
set in the build configuration to allow building as-is from
the official stable release tarball, from a zip file download
from GitHub, from the master branch or any other branch that
includes these changes.

The release script now substitutes the latest stable tag
for the placeholder value in the source/conf.py Sphinx
build configuration file. This modified build configuration
is then included in the release tarball for use by
downstream package maintainers and anyone else who wishes
to build from source using the release tarball.

If the build configuration is NOT modified by the release
script, but used directly from the Git repo, then the
placeholders are intact and serve as a flag to enable dev
builds. These dev builds mirror the prior default
behavior of using verbose date and commit hash values in
addition to also now including the specific branch that
the docs were built from.

Other changes include:

- Update cron script to omit inclusion of build scripts
  in the release tarball. At some point we may include
  a 'contrib' directory for that content. This script
  has also been updated to use the same logic as the
  release_build.sh script for generating the stable
  branch and latest stable tag builds.

- Create small module with various functions that replace
  the unorganized subprocess calls used previously to
  populate various values used by dev builds.

- Update release script to dyanamically calculate values
  from the a Git repo, error out if not run within a repo.
  The default behavior is still to generate a release
  tarball, but also to generate a release Sphinx build
  configuration file for use by downstream package
  maintainers and anyone else who wishes to build from
  source using the release tarball.

- A a new BUILDS_README has been included to cover basic steps
  for building dev and release versions of the docs. This doc
  is intended for rsyslog-doc team members and regular
  contributors whereas the main README is intended for
  occasional contributors or newcomers to the project.

refs rsyslog/rsyslog-doc#504, rsyslog/rsyslog-doc#424