Remove broken footnotes, 'Related Links' section
The footnote references were broken and the 'Related Links'
section was bare, so I pruned them and moved the Sponsors link
just below the 'Sponsors and Community' section intentionally
so it stands out.

Once some PRs ahead of the one I will submit for this set of changes
goes through I plan to move the external link reference to
the substitution definitions file (again, in another PR) since
this is probably not the only place we will reference the Sponsors
link.

Additionally, I removed the explicit link text for the "How
You Can Help" page since the title for that page is used if
you do not override it with explicit link text.

refs rsyslog/rsyslog-doc#502