Proposed fix for handling journal correctly
The fix is to immediately setup the inotify file descriptor via
`sd_journal_get_fd()` right after a journal open, and then
periodically call `sd_journal_process()` to give the client API
library a chance to detect deleted journal files on disk that need to
be closed so they can be properly erased by the file system.

We remove the open/close dance and simplify that code as a result.

Fixes issue #2436.