Build: Set 'simple' dev build string as default
Prior hard-coded dev build string:

  "rsyslog 8.xx-BRANCH-YYYYMMDD-commit_hash documentation"

That format has been given a name of 'detailed' and
a new simpler format defined as:

  "rsyslog commit_hash docs"

Stable release build strings still use the format of:

  "rsyslog 8.x.x documentation"

As before, epub release builds default the release build
strings. epub dev builds still use the detailed
release string format since the display of epub files
is not affected in the same way as the HTML format is
when using themes that are more narrowly restricted in
horizontal space.

refs rsyslog/rsyslog-doc#537