Fix data alignment problems in plugins/imuxsock/imuxsock.c
gcc would (rightly) complain that the cred and ts pointers
would cause alignment problems, so they were converted to structs
and the necessary data was memcpy()'ed to them.

the aux[] buffer was also potentially misaligned, so making a union
out of it and struct cmsghdr insured it was properly aligned.

The problems were especially visible on ia64 machines:
[6143331.481422] ia64_handle_unaligned: 2 callbacks suppressed
[6143331.481463] in:imuxsock(31387): unaligned access to 0x2000000000e2e554, ip=0x20000000008594c0
[6143331.481519] in:imuxsock(31387): unaligned access to 0x2000000000e2e55c, ip=0x20000000008594e0