testbench: do not run some DB tests twice in split test runs
we can split the testbench into two test runs in order to make each of them complete
quicker (this currently is required for Travis, but may also be useful for other
cases). Some test are run twice in this scenario, which is obviously counter-productive.
We ensure now that at least some longer-running database tests are only executed
once in this setup.