testbench: add testcase for memory leak in stream class
Bug is currently exposed only via imfile. Happens when in readMode 0 a
partial line is read and no more data is present in the file during
that iteration. One partial message is lost in this case.

Note: this just adds the test, but does not activate it. Activation
shall happen after fix is crafted.

see also https://github.com/rsyslog/rsyslog/issues/2421