Link directly to properties documentation
To see a list of properties, one has to follow the link from the Property Replacement page to get to the "rsyslog Properties" page. Might as well just link directly there from here. I imagine this may have happened becasue the rsyslog Properties didn't used to be a separate page?