omkafka: use "atomic mutex macros"
commit 87c8f478ebd0111c9c8d521020acffa42a4508c9 fixed missing mutex
definition, but did not do so via the macros originally intended for
this purpose. It still was very useful, as the patch made omkafka
compile on systems without atomic operations. And did so without
causing any real overhead.

Nevertheless this commit now adjust to use the rsyslog-core intended
way of doing things.