imudp: replace select() by poll()
select was used when epoll was not available. This could cause issues
if during startup fds > 1024 were generated. This is extremely unlikely
and we have never seen any such report in practice. Nevertheless, it is
better to have this fixed.

as a side-activity improve hostname-getaddrinfo-fail.sh test - it looked
even when OK - clarifying things.

closes https://github.com/rsyslog/rsyslog/issues/2615