core bugfix: segfault on queue shutdown
if a ruleset queue is in direct mode, a segfault can occur during
rsyslog shutdown. The root cause is that a direct queue does not
have an associated worker thread pool, but the ruleset destructor
does not anticipate that and tries to destruct the worker thread
pool. It needs to do this itself, as otherwise we get a race
between rulesets on shutdown.

This is a regression from
  https://github.com/rsyslog/rsyslog/commit/3fbd901b3e6300010

closes https://github.com/rsyslog/rsyslog/issues/2480