cosmetic bugfix: valgrind complain in some error conditions
When a "call" script statement was used with a non-existing ruleset AND
debug logginf was active for rainerscript.c, valgrind complained and
one debug message could potentially be incorrect (stating a queue
where non was).

This was caused by not initializing a variable which was probed in
debug mode before it was set. Non-debug mode was not affected.

see also https://github.com/rsyslog/rsyslog/issues/2399#issuecomment-384890873