imudp: use rsyslog message rate-limiter instead of home-grown one
imudp introduced it's own (feature-limited) rate-limiting capability for
message on disallowed senders before we had central rate-limiters
inside rsyslog. Also, that code evolved from running on a single
thread to running on multiple threads, which introduced data races
and so made unreliable.

Now we removed the old rate-limiting capability and depend on the
system rate limiter for internal rsyslog messages.

closes https://github.com/rsyslog/rsyslog/issues/2467