travis: remove "merge master" test
We originally merged master into the PR. At that time, github did not
(reliably?) provide a merged code basis. This is now the case, so the
test is kind of unnecessary. It only did some good if master was updated
after PR submission but before Travis started the test. This is pretty
unlikely, so we remove that test - which also is a relief in regard
to Travis runtime.

see also https://github.com/rsyslog/rsyslog/issues/2800