testbench: move some cleanup to "make clean" target
To support parallel testbench runs, we need to have dynamic work files.
So deleting work files on each test does not work, especially as we can
no longer assume they are "left-overs" from a failed. tes - they may
actually (and quite likely) belong to tests that run in parallel.

So the proper solution is to do via "make clean".

Note that we still delete some files on test init, as not the whole
testbench is sufficiently converted to work without that. Once the
conversion is completed, this part can be removed by another PR.

closes https://github.com/rsyslog/rsyslog/pull/2916 (which it replaces)