shutdown: fix race in thread termination
This prevents new workers from being spawned when the system
is in shutdown state, except when needed to persist queue
data to disk. In all other cases, re-spawning workers brings
the system to an instable state and is not desired. We think
this happens during very late messages being newly generated
(e.g. status message) or received.