omprog: remove "hideEnvironment" parameter as this is a bugfix
discussion confirmed that not passing environment to called program
should be considered a bug and consequently no parameter to control
this behaviour is needed. I just added that parameter to be able
to carry on with testbench work, and this commit removes it. The
param was never part of any official relase.

closes https://github.com/rsyslog/rsyslog/issues/2921