AIX: adding seemingly missing file strippedsymbols.exp
This file is referenced inside IBM's AIX configure.ac code but
cannot be found inside the source. I found it mentioned in some
mail an added it to the source of the project tree.

I do *NOT* know if the content or location is correct nor am
I able to verify if the build succeeds or not.

see also: https://github.com/rsyslog/rsyslog/issues/2597