omprog: document 'useTransactions' flag as experimental
The rsyslog module interface was not designed to support modules that
can act both as transactional and non-transactional according to
a configuration parameter. The old transaction interface (incidentally)
supports this, but it is not supported by the new transaction interface.

Therefore a future change is envisaged
regarding this feature, that will likely break backwards compatibility.