bugfix testbench: kafka test startup was racy
Topic creation was racy: it can happen that the brokers have not
been synced.  This causes follow-on errors and ultimately causes
the test to hang in rsyslog shutdown.

This PR introduces a method to ensure the brokers are acutally
up and running before the rest of the test is carried out. I
assume this fixes the problems we have seen. However, only
practice will show if this really is the case because the
problem happened relatively infrequently (so a couple of good
tests do not necessarily indicate "problem solved").

As such, we close the issue with this commit, but may need
to re-open it if the issue is seen again.

closes https://github.com/rsyslog/rsyslog/issues/3045