bugfix imfile: segfault in ratelimiter
imfile crashes inside rate limit processing, often when log
files are rotated. However, this could occur in any case where
the monitored files was closed by imfile, it rotation is just
the most probable cause for this (moving the file to another
directory or deleting it also can trigger the same issue, for
example). The root cause was invalid sequence of operations.

closes https://github.com/rsyslog/rsyslog/issues/3021